Adding my Alien.
Interesting things it does:
Assesses energy at current tile at position, compares it to energy at
previous position. If the previous position was better, it moves back.
If the current position is better, it moves to a random position other
than the previous one.
Checks the returns from teching and spawning and changes its behavior
accordingly.